+++ domain/blob_service.py
@@ -1,9 +1,11 @@
+from dataclasses import dataclass
 from logging import warning
 import random
 from sqlalchemy.exc import IntegrityError
 
 from data.db.db_engine import transactional
 from data.model.blob import Blob
+from data.model.event_type import EventType
 from data.persistence.blob_reposiotry import (
     get_all_blobs_by_name,
     get_blob_by_id,
@@ -38,6 +40,12 @@ from domain.utils.constants import (
     PRACTICE_EFFECT,
 )
 from domain.utils.sim_time_utils import format_sim_time_short, get_season
+
+
+@dataclass
+class StatMultiplyers:
+    strength: float
+    speed: float
 
 
 @transactional
@@ -85,6 +93,7 @@ def create_blob(session, first_name: str, last_name: str, parent_id: int | None 
     """Create a new blob with random stats and add it to the queue."""
 
     strength = 0.9 + random.random() * 0.2
+    speed = 0.9 + random.random() * 0.2
     learning = 0.5 + 0.5 * random.random()
     current_time = get_sim_time(session)
     queue = get_queue(session)
@@ -101,6 +110,7 @@ def create_blob(session, first_name: str, last_name: str, parent_id: int | None 
                 first_name=first_name,
                 last_name=last_name,
                 strength=strength,
+                speed=speed,
                 learning=learning,
                 integrity=INITIAL_INTEGRITY,
                 born=current_time,
@@ -119,18 +129,18 @@ def update_blobs(session):
     """Update all blobs living in the simulation and yield the progress in percentage."""
 
     current_event = get_current_calendar(session)
-    current_event_league_id = (
-        current_event.league_id if current_event is not None else None
-    )
     blobs = get_all_blobs_by_name(session)
 
     current_time = get_sim_time(session)
 
     modified_blobs = []
     for blob in blobs:
-        multiplyer = 0
-        if blob.league_id == current_event_league_id:
-            multiplyer = COMPETITION_EFFECT
+        multiplyer = StatMultiplyers(strength=0, speed=0)
+        if blob.league_id == current_event.league_id if current_event is not None else None:
+            if current_event.event_type == EventType.ENDURANCE_RACE:
+                multiplyer.speed = COMPETITION_EFFECT
+            else:
+                multiplyer.strength = COMPETITION_EFFECT
         elif blob.money >= MAINTENANCE_COST and random.random() < 0.5:
             blob.money -= MAINTENANCE_COST
             blob.integrity += MAINTENANCE_EFFECT
@@ -139,8 +149,10 @@ def update_blobs(session):
             if activity == ActivityType.LABOUR:
                 blob.money += LABOUR_SALARY
             elif activity == ActivityType.PRACTICE:
-                multiplyer = PRACTICE_EFFECT
-        blob.strength = _update_blob_strength(blob, multiplyer)
+                ratio = random.random()
+                multiplyer.strength = PRACTICE_EFFECT * ratio
+                multiplyer.speed = PRACTICE_EFFECT * (1 - ratio)
+        _update_blob_stats(blob, multiplyer)
         blob.integrity -= 1
         _terminate_blob(blob, current_time)
 
@@ -150,28 +162,31 @@ def update_blobs(session):
 
 
 @transactional
-def update_blob_strength_by_id(blob_id: int, multiplyer: float, session):
+def update_blob_speed_by_id(blob_id: int, multiplyer: float, session):
     blob = get_blob_by_id(session, blob_id)
     if blob:
-        blob.strength = _update_blob_strength(blob, multiplyer, False)
+        blob.speed = _update_stat(blob.speed, multiplyer, blob.learning, blob.integrity, 0)
         save_blob(session, blob)
 
 
-def _update_blob_strength(blob: Blob, multiplyer: float, add_atrophy: bool = True) -> float:
-    """Update the strength of the blob based on the activity multiplyer, its current integrity and learning."""
+def _update_blob_stats(blob: Blob, multiplyer: StatMultiplyers):
+    """Update the stats of the blob based on the activity multiplyer, its current integrity and learning."""
 
     atrophy = 0
-    if blob.integrity < 0.4 and add_atrophy:
+    if blob.integrity < INITIAL_INTEGRITY * 0.4:
         tippingPoint = INITIAL_INTEGRITY * 0.6
         atrophy = -(blob.integrity - tippingPoint) / (
             1.2 * tippingPoint * CYCLES_PER_SEASON
         )
 
-    return (
-        blob.strength
-        - atrophy
-        + multiplyer * blob.learning * (blob.integrity / INITIAL_INTEGRITY)
-    )
+    blob.strength = _update_stat(blob.strength, multiplyer.strength, blob.learning, blob.integrity, atrophy)
+    blob.speed = _update_stat(blob.speed, multiplyer.speed, blob.learning, blob.integrity, atrophy)
+
+
+def _update_stat(stat: float, multiplyer: float, learning: float, integrity: float, atrophy: float) -> float:
+    """Update the stat of the blob based on the activity multiplyer."""
+
+    return stat - atrophy + multiplyer * learning * (integrity / INITIAL_INTEGRITY)
 
 
 @transactional
@@ -193,8 +208,8 @@ def _create_with_name_suggestion(session) -> str | bool:
 
 
 def _terminate_blob(blob: Blob, current_time: int):
-    """Terminate the blob if it's integrity or strength drops to or below zero."""
+    """Terminate the blob if it's integrity or any stat drops to or below zero."""
 
-    if blob.integrity <= 0 or blob.strength <= 0:
+    if blob.integrity <= 0 or blob.strength <= 0 or blob.speed <= 0:
         blob.league_id = None
         blob.terminated = current_time