+++ domain/dtos/blob_competitor_dto.py
@@ -6,4 +6,5 @@ class BlobCompetitorDto():
     id: int
     name: str
     strength: float
+    speed: float
     color: str