+++ domain/utils/action_utils.py
@@ -8,8 +8,8 @@ def generate_race_score_for_contender(contender: BlobCompetitorDto, current_time
     """
     Generate race score for contender.
     """
-    noise = perlin_like_noise_1d(race_duration, current_time * contender.strength)
-    return noise[tick] * contender.strength
+    noise = perlin_like_noise_1d(race_duration, current_time * contender.speed)
+    return noise[tick] * contender.speed
 
 
 @lru_cache(maxsize=128)